修复帖子相关度一样的时候未按照发帖时间排序的问题

+++ app/Models/User.php
@@ -74,7 +74,7 @@ class User extends BaseModel implements AuthenticatableContract, CanResetPasswor
             'search_analyzer' => 'ik_max_word',
         ],
         'score' => [
-            'type' => 'float'
+            'type' => 'integer'
         ]
     ];
